Rename LCHT to LabelCombinationHoeffdingTree
