+++ src/skmultiflow/trees/__init__.py
@@ -8,17 +8,19 @@ from . import split_criterion
 from .hoeffding_tree import HoeffdingTreeClassifier
 from .hoeffding_adaptive_tree import HoeffdingAdaptiveTreeClassifier
 from .extremely_fast_decision_tree import ExtremelyFastDecisionTreeClassifier
-from .lc_hoeffding_tree import LCHT
+from .label_combination_hoeffding_tree import LabelCombinationHoeffdingTree
 from .regression_hoeffding_tree import RegressionHoeffdingTree
 from .regression_hoeffding_adaptive_tree import RegressionHAT
 from .multi_target_regression_hoeffding_tree import MultiTargetRegressionHoeffdingTree
 from .stacked_single_target_hoeffding_tree_regressor import StackedSingleTargetHoeffdingTreeRegressor
 from .hoeffding_tree import HoeffdingTree   # remove in v0.7.0
 from .hoeffding_adaptive_tree import HAT   # remove in v0.7.0
-from .extremely_fast_decision_tree import HATT   # remove in v0.7.0
+from .extremely_fast_decision_tree import HATT   # remove in v0.7.0"
+from .label_combination_hoeffding_tree import LCHT   # remove in v0.7.0"
 
 __all__ = ["attribute_observer", "attribute_test", "nodes", "split_criterion",
            "HoeffdingTreeClassifier", "HoeffdingAdaptiveTreeClassifier", "ExtremelyFastDecisionTreeClassifier",
+           "LabelCombinationHoeffdingTree",
            "HoeffdingTree", "HAT", "HATT", "LCHT", "RegressionHoeffdingTree",
            "RegressionHAT", "MultiTargetRegressionHoeffdingTree",
            "StackedSingleTargetHoeffdingTreeRegressor"]
