+++ tests/trees/test_label_combination_hoeffding_tree.py
@@ -1,14 +1,14 @@
 import numpy as np
-from skmultiflow.trees import LCHT
+from skmultiflow.trees import LabelCombinationHoeffdingTree
 from skmultiflow.data import MultilabelGenerator
 
 
-def test_lc_hoeffding_tree_mc(test_path):
+def test_label_combination_hoeffding_tree_mc(test_path):
     stream = MultilabelGenerator(n_samples=10000, n_features=15, n_targets=3, n_labels=4, random_state=112)
 
     stream.prepare_for_use()
 
-    learner = LCHT(n_labels=3, leaf_prediction='mc')
+    learner = LabelCombinationHoeffdingTree(n_labels=3, leaf_prediction='mc')
 
     cnt = 0
     max_samples = 5000
@@ -41,21 +41,20 @@ def test_lc_hoeffding_tree_mc(test_path):
     assert type(learner.predict(X)) == np.ndarray
     assert type(learner.predict_proba(X)) == np.ndarray
 
-    expected_info = "LCHT(binary_split=False, grace_period=200, leaf_prediction='mc',\n" \
-                    "     max_byte_size=33554432, memory_estimate_period=1000000, n_labels=3,\n" \
-                    "     nb_threshold=0, no_preprune=False, nominal_attributes=None,\n" \
-                    "     remove_poor_atts=False, split_confidence=1e-07,\n" \
-                    "     split_criterion='info_gain', stop_mem_management=False,\n" \
-                    "     tie_threshold=0.05)"
-    assert learner.get_info() == expected_info
+    expected_info = "LabelCombinationHoeffdingTree(binary_split=False, grace_period=200, leaf_prediction='mc', " \
+                    "max_byte_size=33554432, memory_estimate_period=1000000, n_labels=3, nb_threshold=0, " \
+                    "no_preprune=False, nominal_attributes=None, remove_poor_atts=False, split_confidence=1e-07, " \
+                    "split_criterion='info_gain', stop_mem_management=False, tie_threshold=0.05)"
+    info = " ".join([line.strip() for line in learner.get_info().split()])
+    assert info == expected_info
 
 
-def test_lc_hoeffding_tree_nb(test_path):
+def test_label_combination_hoeffding_tree_nb(test_path):
     stream = MultilabelGenerator(n_samples=10000, n_features=15, n_targets=3, n_labels=4, random_state=112)
 
     stream.prepare_for_use()
 
-    learner = LCHT(n_labels=3, leaf_prediction='nb')
+    learner = LabelCombinationHoeffdingTree(n_labels=3, leaf_prediction='nb')
 
     cnt = 0
     max_samples = 5000
@@ -87,21 +86,20 @@ def test_lc_hoeffding_tree_nb(test_path):
     assert type(learner.predict(X)) == np.ndarray
     assert type(learner.predict_proba(X)) == np.ndarray
 
-    expected_info = "LCHT(binary_split=False, grace_period=200, leaf_prediction='nb',\n" \
-                    "     max_byte_size=33554432, memory_estimate_period=1000000, n_labels=3,\n" \
-                    "     nb_threshold=0, no_preprune=False, nominal_attributes=None,\n" \
-                    "     remove_poor_atts=False, split_confidence=1e-07,\n" \
-                    "     split_criterion='info_gain', stop_mem_management=False,\n" \
-                    "     tie_threshold=0.05)"
-    assert learner.get_info() == expected_info
+    expected_info = "LabelCombinationHoeffdingTree(binary_split=False, grace_period=200, leaf_prediction='nb', " \
+                    "max_byte_size=33554432, memory_estimate_period=1000000, n_labels=3, nb_threshold=0, " \
+                    "no_preprune=False, nominal_attributes=None, remove_poor_atts=False, split_confidence=1e-07, " \
+                    "split_criterion='info_gain', stop_mem_management=False, tie_threshold=0.05)"
+    info = " ".join([line.strip() for line in learner.get_info().split()])
+    assert info == expected_info
 
 
-def test_lc_hoeffding_tree_nba(test_path):
+def test_label_combination_hoeffding_tree_nba(test_path):
     stream = MultilabelGenerator(n_samples=10000, n_features=15, n_targets=3, n_labels=4, random_state=112)
 
     stream.prepare_for_use()
 
-    learner = LCHT(n_labels=3)
+    learner = LabelCombinationHoeffdingTree(n_labels=3)
 
     cnt = 0
     max_samples = 5000
@@ -131,10 +129,9 @@ def test_lc_hoeffding_tree_nba(test_path):
     assert type(learner.predict(X)) == np.ndarray
     assert type(learner.predict_proba(X)) == np.ndarray
 
-    expected_info = "LCHT(binary_split=False, grace_period=200, leaf_prediction='nba',\n" \
-                    "     max_byte_size=33554432, memory_estimate_period=1000000, n_labels=3,\n" \
-                    "     nb_threshold=0, no_preprune=False, nominal_attributes=None,\n" \
-                    "     remove_poor_atts=False, split_confidence=1e-07,\n" \
-                    "     split_criterion='info_gain', stop_mem_management=False,\n" \
-                    "     tie_threshold=0.05)"
-    assert learner.get_info() == expected_info
+    expected_info = "LabelCombinationHoeffdingTree(binary_split=False, grace_period=200, leaf_prediction='nba', " \
+                    "max_byte_size=33554432, memory_estimate_period=1000000, n_labels=3, nb_threshold=0, " \
+                    "no_preprune=False, nominal_attributes=None, remove_poor_atts=False, split_confidence=1e-07, " \
+                    "split_criterion='info_gain', stop_mem_management=False, tie_threshold=0.05)"
+    info = " ".join([line.strip() for line in learner.get_info().split()])
+    assert info == expected_info
